+++ packages/viz/src/components/settings/Diabeloop.js
@@ -28,6 +28,10 @@ import HistoryTable from './DiabeloopHistoryParameters';
 
 import * as datetime from '../../utils/datetime';
 import * as dblData from '../../utils/settings/diabeloopData';
+import PumpTable from './PumpTable';
+import CgmTable from './CgmTable';
+import TerminalTable from './TerminalTable';
+
 import { deviceName, getDeviceMeta } from '../../utils/settings/data';
 
 import styles from './Diabeloop.css';
@@ -97,11 +101,9 @@ const Diabeloop = (props) => {
 
   const parameters = _.get(pumpSettings, 'payload.parameters', null);
   const device = _.get(pumpSettings, 'payload.device', null);
+  const pump = _.get(pumpSettings, 'payload.pump', null);
+  const cgm = _.get(pumpSettings, 'payload.cgm', null);
   const history = _.sortBy(_.cloneDeep(_.get(pumpSettings, 'payload.history', null)), ['changeDate']);
-
-  if (parameters === null || device === null) {
-    return null;
-  }
 
   const parametersByLevel = dblData.getParametersByLevel(parameters);
 
@@ -123,17 +125,9 @@ const Diabeloop = (props) => {
       <div className={styles.settingsContainer}>
         <div className={styles.categoryContainer}>
           <div className={styles.categoryTitle}>{t('Device')}</div>
-          <table className={styles.settingsTable}>
-            <caption className={styles.bdlgSettingsHeader}>
-              {t('Device')}<span className={styles.secondaryLabelWithMain}>{device.name}</span>
-            </caption>
-            <tbody>
-              <tr><td>{t('Manufacturer')}</td><td>{device.manufacturer}</td></tr>
-              <tr><td>{t('Identifier')}</td><td>{device.deviceId}</td></tr>
-              <tr><td>{t('IMEI')}</td><td>{device.imei}</td></tr>
-              <tr><td>{t('Software version')}</td><td>{device.swVersion}</td></tr>
-            </tbody>
-          </table>
+          <TerminalTable device={device} />
+          <PumpTable pump={pump} timePrefs={timePrefs} />
+          <CgmTable cgm={cgm} timePrefs={timePrefs} />
         </div>
         {renderDiabeloopParameters(parametersByLevel)}
         <div className={styles.categoryContainer}>
@@ -166,6 +160,7 @@ Diabeloop.propTypes = {
   deviceKey: PropTypes.oneOf(['diabeloop']).isRequired,
   pumpSettings: PropTypes.shape({
     deviceId: PropTypes.string.isRequired,
+    deviceTime: PropTypes.string.isRequired,
     payload: PropTypes.shape({
       device: PropTypes.shape({
         deviceId: PropTypes.string.isRequired,
@@ -203,5 +198,6 @@ Diabeloop.propTypes = {
     timezoneName: PropTypes.string,
   }).isRequired,
   user: PropTypes.object.isRequired,
+  handleClickHistory: PropTypes.func.isRequired,
 };
 export default Diabeloop;