+++ CHANGELOG.md
@@ -18,6 +18,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-273 Translate units in DBL settings table and display depending on patient profile
 - YLP-299 Change "parameters" for "settings" in English
 - YLP-313 Create dedicated key for bolus graph title
+- YLP-324 Pump and CGM information in DBL settings
 ### Engineering Use
 - YLP-211 Rework Cloudfront deployment services to use an alternate domain name
 - YLP-147 Merge blip dependencies